fix(mcp): host capability gate for inputs, copilot-cli repo scope, test hygiene
B3 - inputs and ${input:id} are a VS Code Copilot feature. Other hosts receive the
placeholder as a literal value, so a server installed into a Kiro or Windsurf
config was written successfully and then failed at startup with no indication why.
Note the references live inside the server definitions (command/args/env/headers),
not just in the inputs array, so dropping the array is not a fix - it leaves
dangling placeholders.

Adds an optional supportsInputs flag to McpLayoutConfig (default false, set only
for the VS Code entries) and rejects an install whose servers reference an input
the host cannot resolve, naming the input ids. Servers with no input references
install unchanged. Reuses collectInputReferences, now public, for detection.
This converts a silent runtime failure into an actionable install-time error.

M7 - copilot-cli had no repository-scope entry, so repo-scope installs failed.
Copilot CLI reads .mcp.json from the project root (VS Code's .vscode/mcp.json
support was removed), so the entry is ${workspaceRoot}/.mcp.json.

D3 - consistent error contract in the locator. getMcpConfigLocation is now the
single resolution entry point and getMcpConfigPath delegates to it; undefined is
returned only for the two expected cases (host declares no file at this scope, or
repository scope with no workspace open), while a malformed template still throws.
Removed getMcpServersKey, which was unused and defaulted to 'servers' for unknown
hosts where the sibling method returned undefined for the same input.

D4 - readWorkspaceMcpConfig and writeWorkspaceMcpConfig resolved the layout twice
and probed the filesystem on each pass. Both now resolve once.

D1 (partial) - removed the no-op conditional re-attachment of tasks/inputs in
serializeMcpConfig. withoutServerMapKeys already copies them, so the guards did
nothing and the comment implied they were load-bearing.

mvgadagi - resolveMcpLayoutConfig now takes TargetType rather than string.

Tests:
- split the overloaded mcp-config-locator-kiro.test.ts, which held path
  resolution, format translation and a service round-trip, into
  mcp-config-locator-scopes, mcp-config-format and mcp-config-service.disk.
- replaced the Date.now() temp directory with fs.mkdtemp; two runs in the same
  millisecond shared a directory.
- teardown now restores vscode.env and the workspace-folders global through a
  finally chain; a throw mid-teardown previously left host detection pointing at
  Kiro for every later suite in the run.
- new coverage for the repository-scope refusal on hosts with no workspace-level
  file, the git-exclude pattern using forward slashes and the host's real
  filename, vscode-insiders matching vscode, copilot-cli repo scope, and which
  hosts declare input support.

Note found while testing: host detection only recognises kiro, windsurf/devin and
insiders, everything else falling back to vscode. claude-code and copilot-cli are
therefore unreachable from the extension today; their entries are exercised by
unit tests via an explicit host argument and would be used by the CLI.

Verified: build clean; extension 2195 passing / 0 failing; core 228 passing;
lint 0 errors across core, app, infra and the extension. Pre-existing unrelated
failures (15 app, 74 infra) unchanged.
