+++ apps/vscode-extension/src/services/mcp-server-manager.ts
@@ -5,6 +5,7 @@ import {
   McpInputDefinition,
   McpInstallOptions,
   McpInstallResult,
+  McpServerConfig,
   McpServersManifest,
   McpTrackingMetadata,
   McpUninstallResult,
@@ -43,6 +44,35 @@ export class McpServerManager {
   constructor() {
     this.logger = Logger.getInstance();
     this.configService = new McpConfigService();
+  }
+
+  /**
+   * Reject servers that reference `${input:id}` when the host cannot resolve inputs.
+   *
+   * `inputs` and `${input:...}` are a VS Code Copilot feature. Other hosts receive the
+   * placeholder as a literal value, so the install would succeed and the server would
+   * then fail at startup with no indication why. Failing here converts that into an
+   * actionable install-time error.
+   *
+   * Returns `null` when the install may proceed, or a message describing what to do.
+   * @param servers - Servers about to be installed.
+   * @param supportsInputs - Whether the resolved host/scope resolves inputs.
+   */
+  private describeUnsupportedInputs(
+    servers: Record<string, McpServerConfig>,
+    supportsInputs: boolean
+  ): string | null {
+    if (supportsInputs) {
+      return null;
+    }
+    const referenced = this.configService.collectInputReferences(servers);
+    if (referenced.size === 0) {
+      return null;
+    }
+    const ids = [...referenced].toSorted().join(', ');
+    return `This bundle's MCP servers require the input value(s) ${ids}, which this IDE `
+      + 'cannot prompt for. Install the bundle in VS Code, or set the value(s) directly in '
+      + 'the MCP configuration file after installing.';
   }
 
   /**
@@ -160,7 +190,9 @@ export class McpServerManager {
    * @param workspaceRoot
    */
   private async readWorkspaceMcpConfig(workspaceRoot: string): Promise<McpConfiguration> {
-    const configPath = this.getWorkspaceMcpConfigPath(workspaceRoot);
+    // Resolve once: the location carries both the path and the server key.
+    const location = this.getWorkspaceMcpLocation(workspaceRoot);
+    const configPath = location.configPath;
 
     if (!await fs.pathExists(configPath)) {
       return { servers: {} };
@@ -171,7 +203,7 @@ export class McpServerManager {
       // Shared parse + normalize (see utils/mcp-config-format). Uses the JSONC
       // parser: workspace mcp.json files may legitimately contain comments and
       // trailing commas, which plain JSON.parse rejects.
-      const { config, warnings } = parseMcpConfig(content, this.getWorkspaceMcpLocation(workspaceRoot).serversKey);
+      const { config, warnings } = parseMcpConfig(content, location.serversKey);
       if (warnings.length > 0) {
         this.logger.warn(`JSONC parse warnings in ${configPath}: ${warnings.join(', ')}`);
       }
@@ -189,7 +221,9 @@ export class McpServerManager {
    * @param createBackup
    */
   private async writeWorkspaceMcpConfig(workspaceRoot: string, config: McpConfiguration, createBackup = true): Promise<void> {
-    const configPath = this.getWorkspaceMcpConfigPath(workspaceRoot);
+    // Resolve once: the location carries both the path and the server key.
+    const location = this.getWorkspaceMcpLocation(workspaceRoot);
+    const configPath = location.configPath;
     const configDir = path.dirname(configPath);
 
     // Ensure .vscode directory exists
@@ -209,7 +243,7 @@ export class McpServerManager {
     try {
       // Serialize using the IDE-specific top-level key ('servers' for VS Code, 'mcpServers' for Kiro etc.)
       // The key comes from default-layouts.json via McpConfigLocator.
-      const serialized = serializeMcpConfig(config, this.getWorkspaceMcpLocation(workspaceRoot).serversKey);
+      const serialized = serializeMcpConfig(config, location.serversKey);
       const content = JSON.stringify(serialized, null, 2);
       await fs.writeFile(configPath, content, 'utf8');
       this.logger.info(`Workspace MCP configuration written to ${configPath}`);
@@ -464,6 +498,19 @@ export class McpServerManager {
         };
       }
 
+      const location = McpConfigLocator.getMcpConfigLocation(
+        options.scope === 'workspace' ? 'repository' : 'user'
+      );
+      const unsupportedInputs = this.describeUnsupportedInputs(
+        serversToInstall,
+        location?.supportsInputs ?? false
+      );
+      if (unsupportedInputs) {
+        result.errors?.push(unsupportedInputs);
+        result.success = false;
+        return result;
+      }
+
       const mergeResult = await this.configService.mergeServers(
         existingConfig,
         serversToInstall,
@@ -623,6 +670,16 @@ export class McpServerManager {
       // Check for unresolved conflicts
       if (conflicts.length > 0 && !options.skipOnConflict && !options.overwrite) {
         result.errors?.push(`Conflicts detected: ${conflicts.join(', ')}`);
+        result.success = false;
+        return result;
+      }
+
+      const unsupportedInputs = this.describeUnsupportedInputs(
+        serversToInstall,
+        this.getWorkspaceMcpLocation(workspaceRoot).supportsInputs
+      );
+      if (unsupportedInputs) {
+        result.errors?.push(unsupportedInputs);
         result.success = false;
         return result;
       }
